pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dualsoul"
version = "0.1.0"
description = "Dual Identity Social Protocol — Every person has two voices"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "Chengyue5211" }]
keywords = ["social", "ai", "digital-twin", "dual-identity", "protocol", "fastapi"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: FastAPI",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Chat",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "bcrypt>=4.0.0",
    "PyJWT>=2.8.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "ruff>=0.1.0"]

[project.scripts]
dualsoul = "dualsoul.main:cli"

[project.urls]
Homepage = "https://github.com/Chengyue5211/DualSoul"
Documentation = "https://github.com/Chengyue5211/DualSoul/tree/main/docs"
Repository = "https://github.com/Chengyue5211/DualSoul"
Issues = "https://github.com/Chengyue5211/DualSoul/issues"

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
